mappollard: add InitWithStump
For MapPollard, if you want to initialize based on a specific height,
you can't just set the roots and the numleaves. You also need access
to the internal nodes maps in order for Modify to function correctly.

This new helper function allows an external caller to init a MapPollard
properly from a Stump.